end of section 18

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
 import { MessageService } from '../_services/message.service';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -24,7 +25,7 @@ export class MessagesComponent implements OnInit {
 
   loadMessages() {
     this.loading = true;
-    this.messageService.getMessages(this.pageNumber, this.pageSize, this.container).subscribe(response => {
+    this.messageService.getMessages(this.pageNumber, this.pageSize, this.container).subscribe((response: { result: Message[]; pagination: Pagination; }) => {
       this.messages = response.result;
       this.pagination = response.pagination;
       this.loading = false;
@@ -32,14 +33,17 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe((result: any) => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        })
+      }
     })
   }
+    pageChanged(event: any) {
+      this.pageNumber = event.page;
+      this.loadMessages();
+    }
 
-  pageChanged(event: any) {
-    this.pageNumber = event.page;
-    this.loadMessages();
   }
-
-}
